REFACTOR: modularize enemy AI
